requirements.txt
# JARVIS AI Assistant - Voice Features (Browser-based)
# Core AI and Machine Learning
openai>=1.97.0
numpy>=1.24.0
pandas>=2.0.0

# Web Framework and Real-time Communication
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
requests>=2.31.0

# System Monitoring and Utilities
psutil>=5.9.0
python-dotenv>=1.0.0

# Natural Language Processing
nltk>=3.8.0
textblob>=0.17.0

# Web Scraping and APIs
beautifulsoup4>=4.12.0
newsapi-python>=0.2.6

# Additional Utilities
python-dateutil>=2.8.0
pytz>=2023.3
schedule>=1.2.0
click>=8.1.0
rich>=13.0.0
colorama>=0.4.6

# Configuration
pyyaml>=6.0

# Voice Recognition and TTS (for jarvis.py)
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11

# Voice Features: Uses browser's Web Speech API
# - Speech Recognition: Browser's built-in microphone
# - Text-to-Speech: Browser's built-in speakers
# - No system audio dependencies needed! 